Fusion-AI Chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
gTTS
Pillow
pyperclip
pyttsx3
httpx==0.23.3
httpcore==0.16.1
translate
